requirements.txt
customtkinter
Pillow
requests
validators
pyperclip
pygments
pyinstaller
matplotlib